requirements.txt
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
Pillow==10.1.0
torch
open_clip_torch
tensorflow==2.15.0
requests==2.31.0
python-multipart==0.0.6
numpy==1.24.3
 